pyproject.toml
[tool.poetry]
name = "referal-service"
version = "0.1.0"
description = ""
authors = ["Andrey Tarakanov <andrey.t-krd@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
drf-spectacular = "^0.27.1"
flake8 = "^7.0.0"
aiohttp = "^3.9.3"
httpx = "^0.26.0"
celery = "^5.3.6"
redis = "^5.0.1"
adrf = "^0.1.4"
drfasyncview = "^0.2.3"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
